Include files from trash
