+++ web/apps/photos/src/pages/gallery/index.tsx
@@ -1,8 +1,12 @@
 import { WhatsNew } from "@/new/photos/components/WhatsNew";
 import { shouldShowWhatsNew } from "@/new/photos/services/changelog";
 import { fetchAndSaveFeatureFlagsIfNeeded } from "@/new/photos/services/feature-flags";
-import { getLocalFiles } from "@/new/photos/services/files";
+import {
+    getLocalFiles,
+    getLocalTrashedFiles,
+} from "@/new/photos/services/files";
 import { EnteFile } from "@/new/photos/types/file";
+import { mergeMetadata } from "@/new/photos/utils/file";
 import log from "@/next/log";
 import { CenteredFlex } from "@ente/shared/components/Container";
 import EnteSpinner from "@ente/shared/components/EnteSpinner";
@@ -94,7 +98,7 @@ import { syncCLIPEmbeddings } from "services/embeddingService";
 import { syncEntities } from "services/entityService";
 import { syncFiles } from "services/fileService";
 import locationSearchService from "services/locationSearchService";
-import { getLocalTrashedFiles, syncTrash } from "services/trashService";
+import { syncTrash } from "services/trashService";
 import uploadManager from "services/upload/uploadManager";
 import { isTokenValid, syncMapEnabled } from "services/userService";
 import { Collection, CollectionSummaries } from "types/collection";
@@ -125,7 +129,6 @@ import {
     getSelectedFiles,
     getUniqueFiles,
     handleFileOps,
-    mergeMetadata,
     sortFiles,
 } from "utils/file";
 import { isArchivedFile } from "utils/magicMetadata";
