+++ web/apps/photos/src/pages/shared-albums/index.tsx
@@ -1,4 +1,5 @@
 import { EnteFile } from "@/new/photos/types/file";
+import { mergeMetadata } from "@/new/photos/utils/file";
 import log from "@/next/log";
 import {
     CenteredFlex,
@@ -65,12 +66,7 @@ import {
     SetFilesDownloadProgressAttributesCreator,
 } from "types/gallery";
 import { downloadCollectionFiles, isHiddenCollection } from "utils/collection";
-import {
-    downloadSelectedFiles,
-    getSelectedFiles,
-    mergeMetadata,
-    sortFiles,
-} from "utils/file";
+import { downloadSelectedFiles, getSelectedFiles, sortFiles } from "utils/file";
 import { PublicCollectionGalleryContext } from "utils/publicCollectionGallery";
 
 export default function PublicCollectionGallery() {
