+++ web/apps/photos/src/services/embeddingService.ts
@@ -1,5 +1,8 @@
 import type { EmbeddingModel } from "@/new/photos/services/embedding";
-import { getAllLocalFiles } from "@/new/photos/services/files";
+import {
+    getAllLocalFiles,
+    getLocalTrashedFiles,
+} from "@/new/photos/services/files";
 import { EnteFile } from "@/new/photos/types/file";
 import { inWorker } from "@/next/env";
 import log from "@/next/log";
@@ -18,7 +21,6 @@ import type {
 } from "types/embedding";
 import { getLocalCollections } from "./collectionService";
 import type { FaceIndex } from "./face/types";
-import { getLocalTrashedFiles } from "./trashService";
 
 type FileML = FaceIndex & {
     updatedAt: number;
