--- web/apps/photos/src/types/trash/index.ts
@@ -1,16 +1,0 @@
-import { EncryptedEnteFile, EnteFile } from "@/new/photos/types/file";
-
-export interface TrashItem extends Omit<EncryptedTrashItem, "file"> {
-    file: EnteFile;
-}
-
-export interface EncryptedTrashItem {
-    file: EncryptedEnteFile;
-    isDeleted: boolean;
-    isRestored: boolean;
-    deleteBy: number;
-    createdAt: number;
-    updatedAt: number;
-}
-
-export type Trash = TrashItem[];
