+++ web/packages/new/photos/services/files.ts
@@ -1,7 +1,8 @@
-import { type EnteFile } from "@/new/photos/types/file";
 import log from "@/next/log";
 import { Events, eventBus } from "@ente/shared/events";
 import localForage from "@ente/shared/storage/localForage";
+import { type EnteFile, type Trash } from "../types/file";
+import { mergeMetadata } from "../utils/file";
 
 const FILES_TABLE = "files";
 const HIDDEN_FILES_TABLE = "hidden-files";
@@ -35,4 +36,42 @@ export const setLocalFiles = async (
     } catch (e) {
         log.error("Failed to save files", e);
     }
+};
+
+export const TRASH = "file-trash";
+
+export async function getLocalTrash() {
+    const trash = (await localForage.getItem<Trash>(TRASH)) ?? [];
+    return trash;
+}
+
+export async function getLocalTrashedFiles() {
+    return getTrashedFiles(await getLocalTrash());
+}
+
+export function getTrashedFiles(trash: Trash): EnteFile[] {
+    return sortTrashFiles(
+        mergeMetadata(
+            trash.map((trashedFile) => ({
+                ...trashedFile.file,
+                updationTime: trashedFile.updatedAt,
+                deleteBy: trashedFile.deleteBy,
+                isTrashed: true,
+            })),
+        ),
+    );
+}
+
+const sortTrashFiles = (files: EnteFile[]) => {
+    return files.sort((a, b) => {
+        if (a.deleteBy === b.deleteBy) {
+            if (a.metadata.creationTime === b.metadata.creationTime) {
+                return (
+                    b.metadata.modificationTime - a.metadata.modificationTime
+                );
+            }
+            return b.metadata.creationTime - a.metadata.creationTime;
+        }
+        return (a.deleteBy ?? 0) - (b.deleteBy ?? 0);
+    });
 };
