+++ web/packages/new/photos/types/file.ts
@@ -126,3 +126,18 @@ export interface FilePublicMagicMetadataProps {
 
 export type FilePublicMagicMetadata =
     MagicMetadataCore<FilePublicMagicMetadataProps>;
+
+export interface TrashItem extends Omit<EncryptedTrashItem, "file"> {
+    file: EnteFile;
+}
+
+export interface EncryptedTrashItem {
+    file: EncryptedEnteFile;
+    isDeleted: boolean;
+    isRestored: boolean;
+    deleteBy: number;
+    createdAt: number;
+    updatedAt: number;
+}
+
+export type Trash = TrashItem[];
